chore: move ChainDropdown into separate directory
