+++ src/components/StakingVaults/DeFiEarn.tsx
@@ -4,7 +4,7 @@ import type { ChainId } from '@shapeshiftoss/caip'
 import type { Property } from 'csstype'
 import { useState } from 'react'
 import { useTranslate } from 'react-polyglot'
-import { ChainDropdown } from 'components/AssetSearch/Chains/ChainDropdown'
+import { ChainDropdown } from 'components/ChainDropdown/ChainDropdown'
 import { useQuery } from 'hooks/useQuery/useQuery'
 import { selectPortfolioChainIdsSortedUserCurrency } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
